+++ src/Kiota.Builder/Configuration/GenerationConfiguration.cs
@@ -107,7 +107,6 @@ public class GenerationConfiguration : ICloneable
     }
     private static readonly HashSet<GenerationLanguage> BarreledLanguages = [
         GenerationLanguage.Ruby,
-        GenerationLanguage.Swift,
     ];
     private static readonly HashSet<GenerationLanguage> BarreledLanguagesWithConstantFileName = [];
     public bool CleanOutput