+++ src/Kiota.Builder/Writers/TypeScript/CodeFunctionWriter.cs
@@ -398,10 +398,11 @@ public class CodeFunctionWriter(TypeScriptConventionService conventionService) :
             } param)
             throw new InvalidOperationException("Interface parameter not found for code interface");
 
-        writer.StartBlock($"if ({param.Name.ToFirstCharacterLowerCase()}) {{");
+        var serializingDerivedTypeParam = codeElement.OriginalLocalMethod.Parameters.FirstOrDefault(static x => x.Kind is CodeParameterKind.SerializingDerivedType) ?? throw new InvalidOperationException("Serializing derived type parameter not found");
+        writer.WriteLine($"if (!{param.Name.ToFirstCharacterLowerCase()} || {serializingDerivedTypeParam.Name}) {{ return; }}");
         if (codeInterface.StartBlock.Implements.FirstOrDefault(static x => x.TypeDefinition is CodeInterface) is CodeType inherits)
         {
-            writer.WriteLine($"serialize{inherits.TypeDefinition!.Name.ToFirstCharacterUpperCase()}(writer, {param.Name.ToFirstCharacterLowerCase()})");
+            writer.WriteLine($"{GetSerializerFunctionName(codeElement, inherits)}(writer, {param.Name.ToFirstCharacterLowerCase()}, {serializingDerivedTypeParam.Name})");
         }
 
         foreach (var otherProp in codeInterface.Properties.Where(static x => x.IsOfKind(CodePropertyKind.Custom) && !x.ExistsInBaseType && !x.ReadOnly))
@@ -411,7 +412,33 @@ public class CodeFunctionWriter(TypeScriptConventionService conventionService) :
 
         if (codeInterface.GetPropertyOfKind(CodePropertyKind.AdditionalData) is CodeProperty additionalDataProperty)
             writer.WriteLine($"writer.writeAdditionalData({codeInterface.Name.ToFirstCharacterLowerCase()}.{additionalDataProperty.Name.ToFirstCharacterLowerCase()});");
-        writer.CloseBlock();
+
+        if (codeElement.OriginalMethodParentClass.DiscriminatorInformation.HasBasicDiscriminatorInformation &&
+            FindDiscriminatorPropertyBySerializationName(codeInterface, codeElement.OriginalMethodParentClass.DiscriminatorInformation.DiscriminatorPropertyName) is { } discriminatorProperty)
+        {
+            writer.StartBlock($"switch ({param.Name.ToFirstCharacterLowerCase()}.{discriminatorProperty.Name.ToFirstCharacterLowerCase()}) {{");
+            foreach (var mapping in codeElement.OriginalMethodParentClass.DiscriminatorInformation.DiscriminatorMappings)
+            {
+                var mappedType = mapping.Value;
+                writer.StartBlock($"case \"{mapping.Key}\":");
+                writer.WriteLine($"{GetSerializerFunctionName(codeElement, mappedType)}(writer, {param.Name.ToFirstCharacterLowerCase()}, true);");
+                writer.CloseBlock("break;");
+            }
+            writer.CloseBlock();
+        }
+    }
+    private static CodeProperty? FindDiscriminatorPropertyBySerializationName(CodeInterface codeInterface, string serializationName)
+    {
+        if (string.IsNullOrEmpty(serializationName)) return null;
+        if (codeInterface.Properties.FirstOrDefault(prop => prop.WireName.EqualsIgnoreCase(serializationName)) is { } discriminatorProperty)
+            return discriminatorProperty;
+        if (codeInterface.StartBlock.Implements.Any())
+            return codeInterface.StartBlock.Implements
+                .Select(x => x.TypeDefinition)
+                .OfType<CodeInterface>()
+                .Select(x => FindDiscriminatorPropertyBySerializationName(x, serializationName))
+                .FirstOrDefault(x => x is not null);
+        return null;
     }
 
     private static bool IsCollectionOfEnum(CodeProperty property)